move building prompt_ids into get_model_respsonse
