+++ verifiers/envs/multiturn_env.py
@@ -1,6 +1,5 @@
 import logging
 from abc import abstractmethod
-from typing import cast
 
 from openai import AsyncOpenAI
 
@@ -18,7 +17,6 @@ from verifiers.utils.response_utils import (
     parse_response_messages,
     parse_response_tokens,
 )
-from verifiers.utils.token_utils import tokenize_vllm
 
 logger = logging.getLogger(__name__)
 
@@ -64,123 +62,6 @@ class MultiTurnEnv(vf.Environment):
             env_response = await self.env_response(messages, state)
             return concat_messages([messages, env_response])
 
-    async def get_prompt_messages_and_ids(
-        self, state: State, client: AsyncOpenAI
-    ) -> tuple[Messages, list[int]]:
-        if len(state["trajectory"]) == 0:
-            logger.warning(
-                "Calling `get_prompt_messages_and_ids` in first turn. This creates unnecessary overhead, and should not happen."
-            )
-            prompt_messages = state["prompt"]
-            prompt_ids = await tokenize_vllm(
-                client=client,
-                messages=state["prompt"],
-                tools=state["oai_tools"],
-                model=state["model"],
-            )
-            return prompt_messages, prompt_ids
-        else:
-            prev_turn_prompt = state["trajectory"][-1]["prompt"]
-            prev_turn_completion = state["trajectory"][-1]["completion"]
-            prev_turn_tokens = state["trajectory"][-1]["tokens"]
-            assert prev_turn_tokens is not None
-            prev_turn_prompt_ids = prev_turn_tokens["prompt_ids"]
-            prev_turn_completion_ids = prev_turn_tokens["completion_ids"]
-            prev_turn_ids = prev_turn_prompt_ids + prev_turn_completion_ids
-
-            messages = concat_messages([prev_turn_prompt, prev_turn_completion])
-            env_response = await self.env_response(messages, state)
-            messages_and_env_response = concat_messages([messages, env_response])
-
-            def compute_suffix_ids(lst: list[int], value: int) -> list[int]:
-                """Returns all tokens after the last occurrence of `value` in `lst`, if any."""
-
-                def find_last_index(lst: list[int], value: int) -> int:
-                    for i in range(len(lst) - 1, -1, -1):
-                        if lst[i] == value:
-                            return i
-                    raise ValueError
-
-                try:
-                    i = find_last_index(lst, value)
-                    suffix_ids = lst[i + 1 :]
-                    return suffix_ids
-                except ValueError:
-                    # end of message token not found, so we don't need to add any suffix tokens
-                    return []
-
-            def find_largest_overlap(a: list[int], b: list[int]) -> int:
-                """Find the largest overlapping sequence between the end of a and beginning of b."""
-                if not a or not b:
-                    return 0
-
-                max_possible = min(len(a), len(b))
-                for overlap_len in reversed(range(1, max_possible + 1)):
-                    a_suffix = a[-overlap_len:]
-                    b_prefix = b[:overlap_len]
-
-                    if a_suffix == b_prefix:
-                        return overlap_len
-
-                return 0
-
-            # we build the env_response_ids using simple tokenization
-            env_response_ids = await tokenize_vllm(
-                client=client,
-                messages=env_response,
-                tools=None,
-                model=state["model"],
-            )
-
-            # we add suffix_ids to prev_turn_ids. suffix_ids are tokens that
-            # are added by the chat template after messages, but not
-            # generated by the model, i.e. they will be part of messages_ids
-            # (from the chat template) but not of prev_turn_ids (from the
-            # engine). to not train OOD w.r.t. the chat template, we add
-            # these suffix tokens to prev_turn_ids. we compute the
-            # suffix_ids once, and cache them for future use in 'non-exact'
-            # tokenization mode. then, for each turn, we find the largest
-            # overlap between the end of prev_turn_ids and the beginning of
-            # the suffix_ids. this is to correctly handle truncated turns
-            # that did not produce message delimiting tokens.
-            if getattr(self, "cached_suffix_ids", None) is None:
-                dummy_content = "World!"
-                dummy_messages = cast(
-                    vf.Messages,
-                    [
-                        {"role": "user", "content": "Hello"},
-                        {"role": "assistant", "content": dummy_content},
-                    ],
-                )
-                dummy_content_ids = await tokenize_vllm(
-                    client=client,
-                    messages=dummy_content,
-                    tools=state["oai_tools"],
-                    model=state["model"],
-                )
-                dummy_messages_ids = await tokenize_vllm(
-                    client=client,
-                    messages=dummy_messages,
-                    tools=state["oai_tools"],
-                    model=state["model"],
-                    extra_kwargs=dict(add_generation_prompt=False),
-                )
-                # these are typically chat template specific tokens, such as
-                # eom tokens, newlines, etc.
-                suffix_ids = compute_suffix_ids(
-                    dummy_messages_ids, dummy_content_ids[-1]
-                )
-                setattr(self, "cached_suffix_ids", suffix_ids)
-            else:
-                suffix_ids = getattr(self, "cached_suffix_ids")
-            overlap_len = find_largest_overlap(prev_turn_ids, suffix_ids)
-            prev_turn_ids += suffix_ids[overlap_len:]
-
-            prompt_messages = messages_and_env_response
-            prompt_ids = prev_turn_ids + env_response_ids
-
-            return prompt_messages, prompt_ids
-
     async def add_model_response(
         self,
         state: State,
@@ -222,17 +103,8 @@ class MultiTurnEnv(vf.Environment):
             state["error"] = e
         while not await self.is_completed(state):
             try:
-                if state["use_token_prompts"] and len(state["trajectory"]) > 0:
-                    (
-                        prompt_messages,
-                        prompt_ids,
-                    ) = await self.get_prompt_messages_and_ids(state, client)
-                    response = await self.get_model_response(
-                        state, prompt_messages, prompt_ids=prompt_ids
-                    )
-                else:
-                    prompt_messages = await self.get_prompt_messages(state)
-                    response = await self.get_model_response(state, prompt_messages)
+                prompt_messages = await self.get_prompt_messages(state)
+                response = await self.get_model_response(state, prompt_messages)
                 await self.add_model_response(state, prompt_messages, response)
             except vf.Error as e:
                 state["error"] = e
